requirements.txt
fastapi
uvicorn
aiofiles
google-genai
google-cloud-aiplatform
